requirements.txt
argon2-cffi==25.1.0
cryptography==45.0.4
pyperclip==1.9.0
textual==3.5.0